refactor(cli): split cli.py into a command package with shared helpers (#333)
Restructure the 3,284-line chainweaver/cli.py into a chainweaver/cli/ package:

- _shared.py holds the Typer app/sub-apps, registry state, and the shared
  flow/result loading, error-rendering, and output helpers used across
  commands; flow-loading and executor/tool-import logic now live in exactly
  one place.
- One module per command group (inspect, validate, run, profile, diff,
  attest, suggest, flows, record, doctor, traces, fuzz, service), each
  registering on the shared app at import time and staying well under the
  ~500-line guideline.
- __init__.py wires the submodules, defines the main console-script entry
  point, and re-exports the stable surface (app, set_default_registry,
  ServeTransport, serve_command, _build_flow_server, _import_tools_from,
  profile_command) that host apps and tests rely on.

Behavior is preserved: command bodies are unchanged. Two serve tests that
monkeypatched private helpers on the cli package now target the owning
submodule (chainweaver.cli.run), per the issue's migration note about
private-helper imports.

ruff check, ruff format --check, and mypy --strict pass on the package.
